charset picker part 2, tile palette refactor
in which i decide making a custom model is a waste of time and play pin a button on the dialog instead.

i also moved the tile overview rendering code from the palette to core, as it's useful in a few other places, such as the "upper tile as charset" picker and (not implemented yet) the "replace tiles" command.

this move came with a complete refactor of the tile palette. it now uses a scale on the view instead of manually scaling the tiles and no longer has that extremely cursed "eraser" (this is the first of many opinionated tweaks that will befall this code as i go over it).

it might be wise to refactor all tile-based pickers at some point down the line to use a single boilerplate implementation for maintainability. there's a currently unused one, actually; might be worth checking out.