terrain and tileset tab, part 1
most of this is salvaging the terrain ui file; it seems to have been laid out manually instead of with layout, so it blew up. i still need to do it for the chipset widget.

the chipset widget was commented out due to a compilation failure. fortunately it was pretty easy to debug. it was caused by a typo - the class name of the widget is ChipSetWidget, not ChipsetWidget.

still; i'll probably do a huge refactor of internal naming at some point down the line. the terms panorama and background are used interchangably, same with tileset and chipset, and the codebase is littered with donmiguelisms due to a large majority of the code being made before the steam releases. unfortunately qt creator, my tool of choice, lacks a quick mass refactoring option, so it seems like a rather daunting task, but maintainability and clarity is worth the hassle.