terrain and tileset tab, part 3
did some initial work on binding the terrain tab. it actually partially functions!

the tileset tab is going to be trouble, though, due to a small problem: core only keeps one tileset at a time. before we can move on with this it'll need a refactor that splits the renderer from the cache and caches the tilesets properly. might as well knock this out now - we'll need it later for commands like "transfer map".