split rendering from core, part 3
maps work again!! we're officially back in business. i ended up scrapping the tileset cache because i got too many weird memory issues with it, guess i'll revisit it later.

this also fixes a small bug with how the transparent colors on images are calculated. qpixmaps have premultiplied alpha, and the image loader loaded directly into a pixmap, then converted it to a qimage to check the color. this worked fine for most cases, but unfortunately aseprite devs added a brilliant feature that makes the first colour in your palette transparent even if it's the only layer unless you set it as a background. this sets it to a full black quare as far as a qpixmap is concerned, so if there were any other black pixels in the map they'd be treated as transparent. moving the mask check into the loader and loading a qimage directly to get the values fixed the bug. there seems to be one more use of this value that relies on the check still being done in core, but i've yet to see if it's actually necessary (what are e and f tiles??)

getting the tile picker up and running again next.