split rendering from core, part 3.5
...well that was easy. that only leaves rendering the tile preview in the event info widget and the event graphic picker widget. i'll probably end up splitting the on-map event version of the picker to a separate class simply because of the need to pass the underlying renderer's chipset.

i think i'm bottlenecking on copying the qmap that stores the tiles out of core. i added chipset sharing between the tile picker and the map to reduce some of it (to be used in the event widgets as well), but it still feels way less snappy than it used to. i might have to move the qmap generation code to the rpgpainter, which i kinda wanted to avoid to keep it somewhat flexible but am realizing now was wholly unnecessary.