split rendering from core, part 4
i ended up switching from a ref to a shared_ptr for the shared chipset. this will come in handy later once i add opening more event and map windows at once. what a convenient concept!

anyways, with this we're officially back to the pre-rewrite feature set, which means we can move forward with new features!