only redraw changed tiles
"maple please get back to working on features instead of optimizing"

i'm stumped on the "only redraw necessary parts on scroll" thing because of the way the renderer is structures. there are multiple shifting commands along this chain of commands and i don't know which one's the right one. that one might need a bigger rewrite than i anticipated, so this is probably where i'll call it quits for now. there is a small regression with the rectangle drawing; it's not equipped to clean up after itself. i'll fix this when i add mouse movement interpolation for tile drawing, because the root cause is the same.

this also implements the proper drawing code for the tileset scene widget, except the loops don't work for some reason. i'll have to figure it out soon, since the database widget is next on the agenda. it's looking pretty frightening; i'll have to add more items for the terrain/collision/counter displays and another tile based grid selector, plus there's still the matter of rendering the water tile preview.

i'm kinda tempted to go and work on something a bit easier for once. maybe i'll finish terrain 100%, or maybe i'll poke at that command list rewrite that's been on the horizon, who knows?